refactor code around
